Reorganise Tourism category
